feat: erase invalid input when user requests hint.
Add in features to erase invalid input when user requests hint.
Game now checks for game over condition whenever a hint is requested, or user input is entered, or the solve button is pushed.
If game over condition is met, the game over overlay will be displayed, and the game will be over.
Update TODO.txt to reflect recent changes.